feat: Add ImageCardGrid component for visual problems
Introduces a new component, `ImageCardGrid`, to handle problems that require users to identify parts of images and input corresponding values. This includes updating the `ProblemBody` component to render the new type and adding the `ImageCardGridData` type definition. A sample problem in `topic_1_17.ts` is also refactored to utilize this new component.